+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Next Up
 
+- Elemental Monk 2014 improvements.
+- Improvements to detection of galf/npne damage on save for monster features.
+- Fix some deprecated midiqol effect generations.
+
 # 6.5.11
 
 - When importing with CPR, remove DDB Importer Implementation Details additions