+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Elemental Monk 2014 improvements.
 - Improvements to detection of galf/npne damage on save for monster features.
 - Fix some deprecated midiqol effect generations.
+- Dazzling Footwork Unarmed Strike fixes.
 
 # 6.5.11